requirements.txt
fastapi
uvicorn
python-dotenv
openai
faiss-cpu
numpy
requests
pydantic
langchain
langchain-openai
langchain-community
